implement handler DataBaseException

+++ src/main/java/com/xavier/dscommerce/services/ProductService.java
@@ -3,11 +3,14 @@ package com.xavier.dscommerce.services;
 import com.xavier.dscommerce.Repositories.ProductRepository;
 import com.xavier.dscommerce.dto.ProductDTO;
 import com.xavier.dscommerce.entities.Product;
+import com.xavier.dscommerce.services.exceptions.DataBaseException;
 import com.xavier.dscommerce.services.exceptions.ResouceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -44,9 +47,18 @@ public class ProductService {
         return new ProductDTO(entity);
     }
 
-    @Transactional()
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        productRepository.deleteById(id);
+        if (! productRepository.existsById(id)){
+            throw new ResouceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DataBaseException("Falha de integridade referencial");
+        }
+
     }
     private void copyDtoToEntity(ProductDTO productDTO, Product entity) {
         entity.setName(productDTO.getName());
